restructure platform and web
