+++ src/main/java/sjph/life/platform/text/service/TextCodingHelper.java
